Mend the memory overlap in mem_release_module, ... 5. module /fd32/consdrv.o    [Multiboot-module @ 0x156000, 0x1000 bytes] me = 0x157000 6. module /fd32/winb.o    [Multiboot-module @ 0x157000, 0x8100 bytes] 7. module /command.exe    [Multiboot-module @ 0x166000, 0x24e00 bytes]
If we use the ms1 > me, when we want to release the 5th module
we search for the next module's start, here is 0x157000 the same
as its end, then it will skip to the next module...

Then between its ms and nms will contain two modules and next
time will also free that end of the region. Cause the memory
to free overlapped.


git-svn-id: https://svn.code.sf.net/p/freedos-32/code/trunk@333 85e9c19b-ca11-0410-bd92-afdf7feba87b