Dynalink improvement! The linked object is now able to access the static or global initialized variables.
It's due to the same reason as Nils said. and the detail is that the COMMON_SYMBOLs haven't been relocated in their `destination'.

I've tested with DJGPP and MingW, works fine. Hope that the ELF version can also work.

At the mean time, I also simplified and cleared up a bit.


git-svn-id: https://svn.code.sf.net/p/freedos-32/code/trunk@584 85e9c19b-ca11-0410-bd92-afdf7feba87b